requirements.txt
Faker
fastapi[standard]
python-decouple
python-dotenv
snowplow-tracker